+++ wpull/scraper_test.py
@@ -59,6 +59,7 @@ class TestDocument(unittest.TestCase):
             'http://example.com/object/',  # returned by lxml
             'http://example.com/object/object_data.swf',
             'http://example.com/object/object_archive.dat',
+            'http://example.com/param_ref_value.php',
             'http://example.com/overlay_src.html',
             },
             inline_urls
@@ -275,7 +276,7 @@ class TestDocument(unittest.TestCase):
 
         scrape_info = scraper.scrape(request, response)
 
-        self.assertFalse(scrape_info)
+        self.assertTrue(scrape_info)
 
     def test_scrape_css_urls(self):
         text = '''