+++ wpull/urlfilter.py
@@ -142,6 +142,9 @@ class LevelFilter(BaseURLFilter):
         self._depth = max_depth
 
     def test(self, url_info, url_table_record):
+        if url_table_record.inline:
+            return True
+
         if self._depth:
             return url_table_record.level <= self._depth
         else:
@@ -173,11 +176,14 @@ class ParentFilter(BaseURLFilter):
 
         if schemes_similar(url_info.scheme, top_url_info.scheme) \
         and url_info.hostname == top_url_info.hostname \
-        and url_info.port == top_url_info.port:
+        and (
+            url_info.scheme != top_url_info.scheme or
+            url_info.port == top_url_info.port
+        ):
             return is_subdir(top_url_info.path, url_info.path,
                 trailing_slash=True)
 
-        return False
+        return True
 
 
 class SpanHostsFilter(BaseURLFilter):